# docs/internals/openapi.yaml
openapi: 3.0.3
info:
  title: Cluso GraphDB API
  description: |
    Enterprise-grade graph database API with comprehensive features for managing nodes, edges, and running graph algorithms.
    
    ## Features
    - Node and edge CRUD operations
    - Batch operations for bulk data manipulation
    - Graph traversal and shortest path algorithms
    - Advanced algorithms: PageRank, betweenness centrality, community detection
    - GraphQL and custom query language support
    - JWT and API key authentication
    - Comprehensive audit logging
    - Vector search capabilities (Enterprise edition)
    
    ## Authentication
    This API supports two authentication methods:
    - **JWT Bearer Token**: Use `Authorization: Bearer <token>` header
    - **API Key**: Use `X-API-Key: <key>` header
    
    Most endpoints require authentication. Public endpoints are explicitly marked.
    
  version: 2.0.0
  contact:
    name: GraphDB Support
    email: support@clusographdb.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.graphdb.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /nodes:
    get:
      tags:
        - Nodes
      summary: List nodes
      description: Retrieve a paginated list of nodes with optional filtering
      parameters:
        - name: labels
          in: query
          description: Filter by node labels (comma-separated)
          schema:
            type: string
            example: Person,Company
        - name: limit
          in: query
          description: Maximum number of nodes to return
          schema:
            type: integer
            default: 100
            maximum: 10000
            example: 50
        - name: offset
          in: query
          description: Number of nodes to skip
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: List of nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
                  total:
                    type: integer
                    example: 150000
                  limit:
                    type: integer
                    example: 50
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Nodes
      summary: Create a new node
      description: Create a single node with labels and properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - labels
              properties:
                labels:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  example: ["Person", "Employee"]
                properties:
                  type: object
                  maxProperties: 100
                  additionalProperties: true
                  example:
                    name: Alice Johnson
                    age: 30
                    email: alice@example.com
      responses:
        '201':
          description: Node created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '202':
          description: >-
            Accepted - the node applied in memory but its WAL append
            failed, so it is not durable yet. Do not retry: a retry
            creates the node a second time.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeNotDurableResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: >-
            Service unavailable - a uniqueness rule required for one of the
            node's labels (GRAPHDB_REQUIRED_UNIQUENESS_RULES, ADR 0001) is
            not registered. Fixed wording, naming only the caller's own
            label — never the rule name, never the rest of the required
            list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: required uniqueness rule missing
                  message:
                    type: string
                    example: a required uniqueness rule for label "Claim" is not registered; contact the administrator
                  code:
                    type: integer
                    example: 503
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    Node:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        labels:
          type: array
          items:
            type: string
          example: ["Person", "Employee"]
        properties:
          type: object
          additionalProperties: true
          example:
            name: Alice Johnson
            age: 30
            email: alice@example.com
...
    WriteNotDurableResponse:
      type: object
      description: >-
        Returned with status 202 when a write's change applied in memory
        but its WAL (write-ahead log) append failed. Every reader already
        sees the change, and it becomes durable at the next snapshot or
        clean shutdown. Do not retry: a retry applies the change a second
        time. Returned by updateNode, deleteNode, updateEdge, deleteEdge,
        and the vector index create/delete endpoints, none of which have
        a single per-write entity to return in full.
      properties:
        id:
          type: integer
          format: int64
          description: >-
            The affected entity's id. Omitted (the field is absent) for
            the vector index endpoints, which have no per-write entity id.
          example: 12345
        applied:
          type: boolean
          description: The change applied in memory.
          example: true
        durable:
          type: boolean
          description: The change is not yet durable on disk.
          example: false
        retry:
          type: boolean
          description: Always false. A retry would apply the change a second time.
          example: false
        error:
          type: string
          example: WAL write failed
        message:
          type: string
          example: >-
            The write applied in memory and is not durable yet. Do not
            retry: a retry applies the change a second time. A snapshot or
            a clean shutdown makes it durable.
    NodeNotDurableResponse:
      description: >-
        Returned by POST /nodes with status 202 when the WAL append
        failed. A SUPERSET of the normal 201 Node body — the same id,
        labels and properties — plus the not-durable status fields below,
        so a client that treats any 2xx as success and reads only the node
        fields still builds a correct node.
      allOf:
        - $ref: '#/components/schemas/Node'
        - $ref: '#/components/schemas/WriteNotDurableResponse'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request parameters
        code:
          type: string
          example: INVALID_REQUEST
        details:
          type: object
          additionalProperties: true

    # Vector Search schemas
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request parameters
            code: INVALID_REQUEST
    Unauthorized:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Missing authentication (Bearer token or X-API-Key header required)
            code: UNAUTHORIZED
...