# pkg/api/handlers_nodes.go
package api

import (
	"errors"
	"fmt"
	"log"
	"net/http"
	"strconv"
	"time"

	"github.com/dd0wney/graphdb/pkg/storage"
	"github.com/dd0wney/graphdb/pkg/validation"
)

func (s *Server) handleNodes(w http.ResponseWriter, r *http.Request) {
	s.NewMethodRouter(w, r).
		Get(func() { s.listNodes(w, r) }).
		Head(func() { s.countNodes(w, r) }).
		Post(func() { s.createNode(w, r) }).
		Delete(func() { s.deleteAllNodes(w, r) }).
		NotAllowed()
}

// countNodes responds to HEAD /v1/nodes with the X-Total-Count header
// holding the count of caller-tenant nodes (optionally filtered by
// ?label=). No response body — RFC 9110 §9.3.2 contract.
//
// The unfiltered path uses the O(1) CountNodesForTenant primitive
// (maintained as a counter via increment/decrement on create/delete); the
// filtered path uses CountNodesByLabelForTenant, which reads len(index)
// directly rather than cloning the whole label bucket to take its length
// (audit M1). Neither serializes a response body.
func (s *Server) countNodes(w http.ResponseWriter, r *http.Request) {
	tenantID := getTenantFromContext(r)
	var count uint64
	if label := r.URL.Query().Get("label"); label != "" {
		count = uint64(s.graph.CountNodesByLabelForTenant(tenantID, label))
	} else {
		count = s.graph.CountNodesForTenant(tenantID)
	}
	w.Header().Set("X-Total-Count", strconv.FormatUint(count, 10))
	w.WriteHeader(http.StatusOK)
}

// deleteAllNodes removes all nodes and edges for the CALLER's tenant only.
// Used by single-tenant consumers (e.g. wiki-graph) before a full reload.
// Tenant-scoped (audit/ROADMAP B1): the previous global DeleteAllNodes let any
// authenticated tenant wipe every tenant's data.
//
// Loops over DeleteNodeForTenant directly rather than delegating to the
// storage-level DeleteAllNodesForTenant bulk helper (node_operations.go),
// which mirrors this same loop but ABORTS on the first error. A WAL append
// failure on one node's delete must not stop the sweep of the rest of the
// tenant's nodes — the delete already applied in memory — so this loop
// remembers that and continues, then answers 202 (not durable) once the
// whole tenant has been swept instead of a mid-wipe 500. Any other error
// keeps the pre-existing abort-and-500 path, matching
// DeleteAllNodesForTenant's own contract.
func (s *Server) deleteAllNodes(w http.ResponseWriter, r *http.Request) {
	tenantID := getTenantFromContext(r)

	nodes, enumErr := s.graph.GetAllNodesForTenant(tenantID) // snapshot before mutating
	var walErr error
	for _, n := range nodes {
		if err := s.graph.DeleteNodeForTenant(n.ID, tenantID); err != nil {
			if errors.Is(err, storage.ErrNodeNotFound) {
				continue // concurrent delete already removed it — that's the goal
			}
			if errors.Is(err, storage.ErrWALWriteFailed) {
				// The delete applied in memory; keep sweeping the rest of
				// the tenant's nodes instead of aborting mid-wipe.
				if walErr == nil {
					walErr = err
				}
				continue
			}
			s.respondError(w, http.StatusInternalServerError, sanitizeError(err, "delete all nodes"))
			return
		}
	}
	if enumErr != nil {
		s.respondError(w, http.StatusInternalServerError, sanitizeError(enumErr, "delete all nodes"))
		return
	}
	if walErr != nil {
		s.respondWALWriteFailed(w, walErr, 0)
		return
	}

	s.respondJSON(w, http.StatusOK, map[string]string{"status": "cleared"})
}

func (s *Server) listNodes(w http.ResponseWriter, r *http.Request) {
	// Tenant-scoped per audit Security CRIT #2 (2026-05-06). Previously this
	// called GetAllNodes which dumped every tenant's nodes — any
	// authenticated user could read the full multi-tenant corpus in a single
	// request. The endpoint is registered with requireAuth (server.go:42)
	// but not yet withTenant (audit task A5); getTenantFromContext returns
	// the default tenant when no context is set, which matches existing
	// single-tenant deployment behaviour.
	tenantID := getTenantFromContext(r)

	// Parse pagination first so a malformed ?cursor= / ?limit= surfaces as
	// 400 before we materialize the full tenant list.
	page, status, msg := parsePageRequest(r)
	if status != 0 {
		s.respondError(w, status, msg)
		return
	}

	// Route through the index-level page methods so only the requested page
	// is cloned, not the whole tenant set. Empty ?label= is treated as absent
	// (same as omitting the parameter) — a typo like `?label=` returns zero
	// results rather than the full unfiltered corpus.
	var pageItems []*storage.Node
	var next uint64
	var err error
	if label := r.URL.Query().Get("label"); label != "" {
		pageItems, next, err = s.graph.NodesByLabelPageForTenant(tenantID, label, page.cursor, page.limit)
	} else {
		pageItems, next, err = s.graph.NodesPageForTenant(tenantID, page.cursor, page.limit)
	}
	// A short page returned as a bare 200 would be a lie a client cannot
	// detect, because it looks exactly like a small tenant. The page is served
	// anyway, with X-Enumeration-Incomplete naming how many records would not
	// decode, so the answer is truthful AND the caller keeps both the records
	// that did decode and the cursor to reach the rest (ADR 0003).
	if err != nil && !s.noteIncompleteEnumeration(w, "list nodes", err) {
		// Not an incomplete enumeration: there is no partial page to serve.
		s.respondError(w, http.StatusInternalServerError, sanitizeError(err, "list nodes"))
		return
	}
	writeNextCursor(w, next)

	nodes := make([]*NodeResponse, 0, len(pageItems))
	for _, node := range pageItems {
		nodes = append(nodes, s.nodeToResponse(r.Context(), node))
	}

	s.respondJSON(w, http.StatusOK, nodes)
}

func (s *Server) createNode(w http.ResponseWriter, r *http.Request) {
	var req NodeRequest
	decoder := s.NewRequestDecoder(w, r)
	decoder.DecodeJSON(&req).ValidateNode(&req)
	if decoder.RespondError() {
		return
	}

	// Convert and sanitize properties
	converter := newPropertyConverter()
	props := converter.ConvertAndSanitize(req.Properties, s.convertToValue)

	// Audit A6a: create with the request's tenant context. Without this,
	// every create lands in the default tenant regardless of the caller's
	// JWT claim — re-introducing the cross-tenant write the audit closed
	// at the storage layer.
	tenantID := getTenantFromContext(r)

	// ADR 0001: every create goes through the one uniqueness-rules lookup
	// path, so this handler carries no domain vocabulary of its own — a
	// deployment's REGISTERED rules decide which labels are unique on which
	// property, not a hardcoded "Claim"/"for_task" pair. REST and GraphQL
	// (pkg/graphql/mutations_resolvers.go) call the same storage method, so
	// neither surface can become a way around the other's enforcement.
	node, err := s.graph.CreateNodeWithUniquenessRulesForTenant(tenantID, req.Labels, props)
	if err != nil {
		// R4: a missing required rule is a fixed 503, naming only the
		// caller's own label — never the rule name, never err.Error() from
		// storage (which would name the rule), and never the required list.
		var missing *storage.RequiredRuleMissingError
		if errors.As(err, &missing) {
			s.respondJSON(w, http.StatusServiceUnavailable, ErrorResponse{
				Error: "required uniqueness rule missing",
				Message: fmt.Sprintf(
					"a required uniqueness rule for label %q is not registered; contact the administrator",
					missing.Label,
				),
				Code: http.StatusServiceUnavailable,
			})
			return
		}
		if errors.Is(err, storage.ErrUniqueConstraintViolation) {
			s.respondError(w, http.StatusConflict, err.Error())
			return
		}
		if errors.Is(err, storage.ErrUniquenessRulePropertyMissing) {
			// A caller-supplied request missing a required field is a
			// client error, not a server error.
			s.respondError(w, http.StatusBadRequest, err.Error())
			return
		}
		if errors.Is(err, storage.ErrWALWriteFailed) {
			// The node applied in memory even though the WAL append did
			// not; tell the caller, don't discard it behind a 500.
			// CreateNodeWithUniquenessRulesForTenant routes to one of two
			// storage methods (CreateNodeWithTenant, when no uniqueness
			// rule matches the labels, or CreateNodeWithUniquePropertyForTenant
			// otherwise), and both return the created node alongside a
			// WAL error, so node is non-nil here.
			s.respondNodeWALWriteFailed(r.Context(), w, err, node)
			return
		}
		s.respondError(w, http.StatusInternalServerError, sanitizeError(err, "create node"))
		return
	}

	response := s.nodeToResponse(r.Context(), node)
	s.respondJSON(w, http.StatusCreated, response)
}

func (s *Server) handleNode(w http.ResponseWriter, r *http.Request) {
	// Extract ID from path
	extractor := s.NewPathExtractor(w, r)
	nodeID, ok := extractor.ExtractUint64("/nodes/")
	if !ok {
		return
	}

	s.NewMethodRouter(w, r).
		Get(func() { s.getNode(w, r, nodeID) }).
		Put(func() { s.updateNode(w, r, nodeID) }).
		Delete(func() { s.deleteNode(w, r, nodeID) }).
		NotAllowed()
}

func (s *Server) getNode(w http.ResponseWriter, r *http.Request, nodeID uint64) {
	tenantID := getTenantFromContext(r)
	node, err := s.graph.GetNodeForTenant(nodeID, tenantID)
	if err != nil {
		// ErrRecordUnreadable is a THIRD condition, and it is not about
		// existence: the snapshot directory lists the ID, but the record bytes
		// did not decode. Only the tenant that owns the record can act on that,
		// and only the owner ever sees this error class here.
		//
		// The ordering that makes it safe is NOT in this handler. The record's
		// tenant string lives inside the record that would not decode, so
		// ownership cannot be read from the record body. GetNodeForTenant
		// establishes it from the mmap membership run instead
		// (tenantOwnsUnreadableNode, node_operations.go), BEFORE it chooses
		// which error to return, and it fails closed: no mmap base, no
		// membership section, or a run that does not list the ID all collapse
		// to ErrNodeNotFound. So reaching this branch already means "the
		// caller owns this ID", and a stranger falls through to the 404 below
		// exactly as it did for a record that never existed.
		//
		// 500, not 503. 503 tells the caller to retry later; damaged bytes on
		// disk do not repair themselves, so 503 would state something untrue.
		if errors.Is(err, storage.ErrRecordUnreadable) {
			// The raw error names a byte offset into the snapshot file, and a
			// future decode failure could wrap a cause built from the record's
			// own bytes. Log it for the operator; send a fixed sentence.
			log.Printf("ERROR [get node]: node %d is unreadable for its owning tenant: %v", nodeID, err)
			s.respondError(w, http.StatusInternalServerError, fmt.Sprintf(
				"Node %d exists but its stored record could not be decoded. The data on "+
					"disk is damaged; restore this record from a backup. Retrying will not help.",
				nodeID))
			return
		}
		// ErrNodeNotFound covers both "doesn't exist" and "exists in
		// another tenant" — the unified error is intentional (no
		// existence-leak side channel). 404 is the right status either way.
		s.respondError(w, http.StatusNotFound, "Node not found")
		return
	}

	response := s.nodeToResponse(r.Context(), node)
	s.respondJSON(w, http.StatusOK, response)
}

func (s *Server) updateNode(w http.ResponseWriter, r *http.Request, nodeID uint64) {
	var req NodeRequest
	decoder := s.NewRequestDecoder(w, r)
	decoder.DecodeJSON(&req)
	if decoder.RespondError() {
		return
	}

	// Validate properties if present
	if req.Properties != nil {
		validationReq := validation.NodeRequest{
			Labels:     []string{"Placeholder"}, // Labels not required for update
			Properties: req.Properties,
		}
		if err := validation.ValidateNodeRequest(&validationReq); err != nil {
			s.respondError(w, http.StatusBadRequest, err.Error())
			return
		}
	}

	// Convert and sanitize properties
	converter := newPropertyConverter()
	props := converter.ConvertAndSanitize(req.Properties, s.convertToValue)

	tenantID := getTenantFromContext(r)
	if err := s.graph.UpdateNodeForTenant(nodeID, props, tenantID); err != nil {
		// Cross-tenant update or genuinely-missing node both surface as
		// ErrNodeNotFound — return 404 to avoid an existence-leak side
		// channel. Only true storage errors should 500.
		if errors.Is(err, storage.ErrNodeNotFound) {
			s.respondError(w, http.StatusNotFound, "Node not found")
			return
		}
		if errors.Is(err, storage.ErrWALWriteFailed) {
			// The caller already holds nodeID; the update applied in
			// memory even though the WAL append did not.
			s.respondWALWriteFailed(w, err, nodeID)
			return
		}
		s.respondError(w, http.StatusInternalServerError, sanitizeError(err, "update node"))
		return
	}

	node, err := s.graph.GetNodeForTenant(nodeID, tenantID)
	if err != nil {
		// Update succeeded but couldn't retrieve the updated node — unusual
		// but not fatal. Don't leak the cause; return success-with-id.
		s.respondJSON(w, http.StatusOK, map[string]any{"updated": nodeID})
		return
	}
	response := s.nodeToResponse(r.Context(), node)
	s.respondJSON(w, http.StatusOK, response)
}

func (s *Server) deleteNode(w http.ResponseWriter, r *http.Request, nodeID uint64) {
	tenantID := getTenantFromContext(r)
	if err := s.graph.DeleteNodeForTenant(nodeID, tenantID); err != nil {
		// Cross-tenant or missing → 404 (no existence leak).
		if errors.Is(err, storage.ErrNodeNotFound) {
			s.respondError(w, http.StatusNotFound, "Node not found")
			return
		}
		if errors.Is(err, storage.ErrWALWriteFailed) {
			// The caller already holds nodeID; the delete applied in
			// memory even though the WAL append did not.
			s.respondWALWriteFailed(w, err, nodeID)
			return
		}
		s.respondError(w, http.StatusInternalServerError, sanitizeError(err, "delete node"))
		return
	}

	s.respondJSON(w, http.StatusOK, map[string]any{"deleted": nodeID})
}

func (s *Server) handleBatchNodes(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		s.respondError(w, http.StatusMethodNotAllowed, "Method not allowed")
		return
	}

	var req BatchNodeRequest
	decoder := s.NewRequestDecoder(w, r)
	decoder.DecodeJSON(&req)
	if decoder.RespondError() {
		return
	}

	// Validate batch size
	if err := validation.ValidateBatchSize(len(req.Nodes)); err != nil {
		s.respondError(w, http.StatusBadRequest, err.Error())
		return
	}

	start := time.Now()
	tenantID := getTenantFromContext(r)
	nodes := make([]*NodeResponse, 0, len(req.Nodes))
	converter := newPropertyConverter()

	// #455: track skipped items so callers can tell "dropped" from
	// "never sent" instead of a bare Created count. Index refers to the
	// item's position in req.Nodes (the request), not the response —
	// dropped items never make it into `nodes`, so a response-relative
	// index would be meaningless. This is purely additive: the existing
	// Nodes/Created/Time fields and partial-success behaviour (CC7,
	// consumer_contract_jailgraph_test.go) are untouched.
	var errs []BatchItemError

	for i, nodeReq := range req.Nodes {
		// Validate each node request
		validationReq := validation.NodeRequest{
			Labels:     nodeReq.Labels,
			Properties: nodeReq.Properties,
		}
		if err := validation.ValidateNodeRequest(&validationReq); err != nil {
			errs = append(errs, BatchItemError{Index: i, Error: err.Error()})
			continue // Skip invalid nodes
		}

		// Convert and sanitize properties
		props := converter.ConvertAndSanitize(nodeReq.Properties, s.convertToValue)

		// Audit A6a: scoped create.
		node, err := s.graph.CreateNodeWithTenant(tenantID, nodeReq.Labels, props)
		if err != nil {
			errs = append(errs, BatchItemError{Index: i, Error: err.Error()})
			continue
		}

		nodes = append(nodes, s.nodeToResponse(r.Context(), node))
	}

	response := BatchNodeResponse{
		Nodes:   nodes,
		Created: len(nodes),
		Time:    time.Since(start).String(),
		Failed:  len(errs),
		Errors:  errs,
	}

	s.respondJSON(w, http.StatusCreated, response)
}